+++ packages/core-data/src/entity-provider.js
@@ -5,7 +5,7 @@ import {
 	createContext,
 	useContext,
 	useCallback,
-	useEffect,
+	useMemo,
 } from '@wordpress/element';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { parse, __unstableSerializeAndClean } from '@wordpress/blocks';
@@ -129,7 +129,7 @@ export function useEntityProp( kind, name, prop, _id ) {
 				[ prop ]: newValue,
 			} );
 		},
-		[ kind, name, id, prop ]
+		[ editEntityRecord, kind, name, id, prop ]
 	);
 
 	return [ value, setValue, fullValue ];
@@ -156,12 +156,12 @@ export function useEntityProp( kind, name, prop, _id ) {
 export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 	const providerId = useEntityId( kind, name );
 	const id = _id ?? providerId;
-	const { content, blocks, meta } = useSelect(
+	const { content, editedBlocks, meta } = useSelect(
 		( select ) => {
 			const { getEditedEntityRecord } = select( STORE_NAME );
 			const editedRecord = getEditedEntityRecord( kind, name, id );
 			return {
-				blocks: editedRecord.blocks,
+				editedBlocks: editedRecord.blocks,
 				content: editedRecord.content,
 				meta: editedRecord.meta,
 			};
@@ -171,29 +171,22 @@ export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 	const { __unstableCreateUndoLevel, editEntityRecord } =
 		useDispatch( STORE_NAME );
 
-	useEffect( () => {
-		// Load the blocks from the content if not already in state
-		// Guard against other instances that might have
-		// set content to a function already or the blocks are already in state.
-		if ( content && typeof content !== 'function' && ! blocks ) {
-			const parsedContent = parse( content );
-			editEntityRecord(
-				kind,
-				name,
-				id,
-				{
-					blocks: parsedContent,
-				},
-				{ undoIgnore: true }
-			);
+	const blocks = useMemo( () => {
+		if ( editedBlocks ) {
+			return editedBlocks;
 		}
-	}, [ content ] );
+
+		return content && typeof content !== 'function'
+			? parse( content )
+			: EMPTY_ARRAY;
+	}, [ editedBlocks, content ] );
 
 	const updateFootnotes = useCallback(
 		( _blocks ) => {
-			if ( ! meta ) return;
+			const output = { blocks: _blocks };
+			if ( ! meta ) return output;
 			// If meta.footnotes is empty, it means the meta is not registered.
-			if ( meta.footnotes === undefined ) return {};
+			if ( meta.footnotes === undefined ) return output;
 
 			const { getRichTextValues } = unlock( blockEditorPrivateApis );
 			const _content = getRichTextValues( _blocks ).join( '' ) || '';
@@ -215,7 +208,8 @@ export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 				: [];
 			const currentOrder = footnotes.map( ( fn ) => fn.id );
 
-			if ( currentOrder.join( '' ) === newOrder.join( '' ) ) return;
+			if ( currentOrder.join( '' ) === newOrder.join( '' ) )
+				return output;
 
 			const newFootnotes = newOrder.map(
 				( fnId ) =>
@@ -225,6 +219,71 @@ export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 						content: '',
 					}
 			);
+
+			function updateAttributes( attributes ) {
+				attributes = { ...attributes };
+
+				for ( const key in attributes ) {
+					const value = attributes[ key ];
+
+					if ( Array.isArray( value ) ) {
+						attributes[ key ] = value.map( updateAttributes );
+						continue;
+					}
+
+					if ( typeof value !== 'string' ) {
+						continue;
+					}
+
+					if ( value.indexOf( 'data-fn' ) === -1 ) {
+						continue;
+					}
+
+					// When we store rich text values, this would no longer
+					// require a regex.
+					const regex =
+						/(<sup[^>]+data-fn="([^"]+)"[^>]*><a[^>]*>)[\d*]*<\/a><\/sup>/g;
+
+					attributes[ key ] = value.replace(
+						regex,
+						( match, opening, fnId ) => {
+							const index = newOrder.indexOf( fnId );
+							return `${ opening }${ index + 1 }</a></sup>`;
+						}
+					);
+
+					const compatRegex =
+						/<a[^>]+data-fn="([^"]+)"[^>]*>\*<\/a>/g;
+
+					attributes[ key ] = attributes[ key ].replace(
+						compatRegex,
+						( match, fnId ) => {
+							const index = newOrder.indexOf( fnId );
+							return `<sup data-fn="${ fnId }" class="fn"><a href="#${ fnId }" id="${ fnId }-link">${
+								index + 1
+							}</a></sup>`;
+						}
+					);
+				}
+
+				return attributes;
+			}
+
+			function updateBlocksAttributes( __blocks ) {
+				return __blocks.map( ( block ) => {
+					return {
+						...block,
+						attributes: updateAttributes( block.attributes ),
+						innerBlocks: updateBlocksAttributes(
+							block.innerBlocks
+						),
+					};
+				} );
+			}
+
+			// We need to go through all block attributs deeply and update the
+			// footnote anchor numbering (textContent) to match the new order.
+			const newBlocks = updateBlocksAttributes( _blocks );
 
 			oldFootnotes = {
 				...oldFootnotes,
@@ -241,6 +300,7 @@ export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 					...meta,
 					footnotes: JSON.stringify( newFootnotes ),
 				},
+				blocks: newBlocks,
 			};
 		},
 		[ meta ]
@@ -258,7 +318,6 @@ export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 			// to make sure the edit makes the post dirty and creates
 			// a new undo level.
 			const edits = {
-				blocks: newBlocks,
 				selection,
 				content: ( { blocks: blocksForSerialization = [] } ) =>
 					__unstableSerializeAndClean( blocksForSerialization ),
@@ -282,12 +341,12 @@ export function useEntityBlockEditor( kind, name, { id: _id } = {} ) {
 		( newBlocks, options ) => {
 			const { selection } = options;
 			const footnotesChanges = updateFootnotes( newBlocks );
-			const edits = { blocks: newBlocks, selection, ...footnotesChanges };
+			const edits = { selection, ...footnotesChanges };
 
 			editEntityRecord( kind, name, id, edits, { isCached: true } );
 		},
 		[ kind, name, id, updateFootnotes, editEntityRecord ]
 	);
 
-	return [ blocks ?? EMPTY_ARRAY, onInput, onChange ];
+	return [ blocks, onInput, onChange ];
 }