+++ packages/editor/src/store/test/selectors.js
@@ -75,10 +75,6 @@ selectorNames.forEach( ( name ) => {
 				};
 			},
 
-			isSavingEntityRecord() {
-				return state.saving && state.saving.requesting;
-			},
-
 			getLastEntitySaveError() {
 				const saving = state.saving;
 				const successful = saving && saving.successful;
@@ -1254,7 +1250,7 @@ describe( 'selectors', () => {
 					title: 'sassel',
 				},
 				saving: {
-					requesting: true,
+					pending: true,
 				},
 			};
 
@@ -1320,7 +1316,9 @@ describe( 'selectors', () => {
 								},
 							],
 						},
-						edits: {},
+						edits: {
+							content: () => {},
+						},
 					},
 				},
 				initialEdits: {},
@@ -1403,9 +1401,8 @@ describe( 'selectors', () => {
 				currentPost: {
 					title: 'sassel',
 				},
-				saving: {
-					requesting: true,
-				},
+				postAutosavingLock: {},
+				saving: {},
 				getCurrentUser() {},
 				hasFetchedAutosaves() {
 					return false;
@@ -1434,9 +1431,8 @@ describe( 'selectors', () => {
 				currentPost: {
 					title: 'sassel',
 				},
-				saving: {
-					requesting: true,
-				},
+				postAutosavingLock: {},
+				saving: {},
 				getCurrentUser() {},
 				hasFetchedAutosaves() {
 					return true;
@@ -1597,14 +1593,13 @@ describe( 'selectors', () => {
 			const state = {
 				editor: {
 					present: {
-						blocks: {
-							value: [],
-						},
 						edits: {},
 					},
 				},
 				initialEdits: {},
-				currentPost: {},
+				currentPost: {
+					content: '',
+				},
 			};
 
 			expect( isEditedPostEmpty( state ) ).toBe( true );
@@ -1626,7 +1621,9 @@ describe( 'selectors', () => {
 								},
 							],
 						},
-						edits: {},
+						edits: {
+							content: () => {},
+						},
 					},
 				},
 				initialEdits: {},
@@ -1656,7 +1653,9 @@ describe( 'selectors', () => {
 								},
 							],
 						},
-						edits: {},
+						edits: {
+							content: () => {},
+						},
 					},
 				},
 				initialEdits: {},
@@ -1666,7 +1665,7 @@ describe( 'selectors', () => {
 			expect( isEditedPostEmpty( state ) ).toBe( true );
 		} );
 
-		it( 'should return false if blocks, but empty content edit', () => {
+		it( 'should return true if blocks, but empty content edit', () => {
 			const state = {
 				editor: {
 					present: {
@@ -1693,7 +1692,7 @@ describe( 'selectors', () => {
 				},
 			};
 
-			expect( isEditedPostEmpty( state ) ).toBe( false );
+			expect( isEditedPostEmpty( state ) ).toBe( true );
 		} );
 
 		it( 'should return true if the post has an empty content property', () => {
@@ -1715,7 +1714,7 @@ describe( 'selectors', () => {
 			expect( isEditedPostEmpty( state ) ).toBe( true );
 		} );
 
-		it( 'should return true if edits include a non-empty content property, but blocks are empty', () => {
+		it( 'should return false if edits include a non-empty content property', () => {
 			const state = {
 				editor: {
 					present: {
@@ -1731,7 +1730,7 @@ describe( 'selectors', () => {
 				currentPost: {},
 			};
 
-			expect( isEditedPostEmpty( state ) ).toBe( true );
+			expect( isEditedPostEmpty( state ) ).toBe( false );
 		} );
 
 		it( 'should return true if empty classic block', () => {
@@ -1750,7 +1749,9 @@ describe( 'selectors', () => {
 								},
 							],
 						},
-						edits: {},
+						edits: {
+							content: () => {},
+						},
 					},
 				},
 				initialEdits: {},
@@ -2017,7 +2018,7 @@ describe( 'selectors', () => {
 		it( 'should return true if the post is currently being saved', () => {
 			const state = {
 				saving: {
-					requesting: true,
+					pending: true,
 				},
 			};
 
@@ -2027,7 +2028,7 @@ describe( 'selectors', () => {
 		it( 'should return false if the post is not currently being saved', () => {
 			const state = {
 				saving: {
-					requesting: false,
+					pending: false,
 				},
 			};
 
@@ -2156,6 +2157,7 @@ describe( 'selectors', () => {
 									attributes: {
 										providerNameSlug: 'instagram',
 									},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2178,6 +2180,7 @@ describe( 'selectors', () => {
 									clientId: 567,
 									name: 'core/embed',
 									attributes: {},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2200,11 +2203,13 @@ describe( 'selectors', () => {
 									clientId: 123,
 									name: 'core/image',
 									attributes: {},
+									innerBlocks: [],
 								},
 								{
 									clientId: 456,
 									name: 'core/quote',
 									attributes: {},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2228,6 +2233,7 @@ describe( 'selectors', () => {
 									clientId: 123,
 									name: 'core/image',
 									attributes: {},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2251,6 +2257,7 @@ describe( 'selectors', () => {
 									clientId: 456,
 									name: 'core/quote',
 									attributes: {},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2276,6 +2283,7 @@ describe( 'selectors', () => {
 									attributes: {
 										providerNameSlug: 'youtube',
 									},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2301,6 +2309,7 @@ describe( 'selectors', () => {
 									attributes: {
 										providerNameSlug: 'soundcloud',
 									},
+									innerBlocks: [],
 								},
 							],
 						},
@@ -2324,11 +2333,13 @@ describe( 'selectors', () => {
 									clientId: 456,
 									name: 'core/quote',
 									attributes: {},
+									innerBlocks: [],
 								},
 								{
 									clientId: 789,
 									name: 'core/paragraph',
 									attributes: {},
+									innerBlocks: [],
 								},
 							],
 						},